more fangling
